Move /api/v1/trades to /api/v1/user/{id}/trades
